fix: resolves double loading issue by disabling SSR for home pages
