+++ src/app/home/[tabname]/HomeContent.tsx
@@ -1,0 +1,89 @@
+"use client";
+
+import React, { useEffect, useState } from "react";
+import { useRouter, useParams } from "next/navigation";
+import SpacePage, { SpacePageArgs } from "@/app/(spaces)/SpacePage";
+import { SpaceConfig } from "@/app/(spaces)/Space";
+import TabBar from "@/common/components/organisms/TabBar";
+import {
+  RESOURCES_TAB_CONFIG,
+  NOUNS_TAB_CONFIG,
+  SOCIAL_TAB_CONFIG,
+  GOVERNANCE_TAB_CONFIG,
+  FUNDED_WORKS_TAB_CONFIG,
+} from "./homePageTabsConfig";
+
+const getTabConfig = (tabName: string) => {
+  switch (tabName) {
+    case "Governance":
+      return GOVERNANCE_TAB_CONFIG;
+    case "Resources":
+      return RESOURCES_TAB_CONFIG;
+    case "Funded Works":
+      return FUNDED_WORKS_TAB_CONFIG;
+    case "Social":
+      return SOCIAL_TAB_CONFIG;
+    case "Nouns":
+      return NOUNS_TAB_CONFIG;
+    default:
+      return NOUNS_TAB_CONFIG;
+  }
+};
+
+const HomeContent = () => {
+  const router = useRouter();
+  const params = useParams();
+
+  // Local state to manage current tab name and ordering
+  const tabOrdering = ["Nouns","Social", "Governance", "Resources", "Funded Works"];
+  
+  // Inicializa diretamente com o valor correto para evitar re-renderização
+  const initialTabName = params?.tabname 
+    ? decodeURIComponent(params.tabname as string) 
+    : "Nouns";
+  const [tabName, setTabName] = useState<string>(initialTabName);
+
+  useEffect(() => {
+    const newTabName = params?.tabname
+      ? decodeURIComponent(params.tabname as string)
+      : "Nouns";
+
+    if (newTabName !== tabName) {
+      setTabName(newTabName);
+    }
+  }, [params?.tabname, tabName]);
+
+  function switchTabTo(newTabName: string) {
+    router.push(`/home/${newTabName}`);
+  }
+
+  const tabBar = (
+    <TabBar
+      getSpacePageUrl={(tab) => `/home/${tab}`}
+      inHomebase={false}
+      currentTab={tabName}
+      tabList={tabOrdering}
+      switchTabTo={switchTabTo}
+      inEditMode={false}
+      updateTabOrder={async () => Promise.resolve()}
+      deleteTab={async () => Promise.resolve()}
+      createTab={async () => Promise.resolve({ tabName })}
+      renameTab={async () => Promise.resolve({ tabName })}
+      commitTab={async () => Promise.resolve()}
+      commitTabOrder={async () => Promise.resolve()}
+    />
+  );
+
+  // Sempre usar a configuração real, sem loading/initializing state
+  const args: SpacePageArgs = {
+    config: getTabConfig(tabName) as SpaceConfig,
+    saveConfig: async () => {},
+    commitConfig: async () => {},
+    resetConfig: async () => {},
+    tabBar: tabBar,
+  };
+
+  return <SpacePage {...args} />;
+};
+
+export default HomeContent;
