+++ src/app/home/[tabname]/page.tsx
@@ -1,105 +1,3 @@
-"use client";
+import HomeWrapper from './HomeWrapper';
 
-import React, { useEffect, useState } from "react";
-import { useRouter, useParams } from "next/navigation";
-import { useAppStore } from "@/common/data/stores/app";
-import SpacePage, { SpacePageArgs } from "@/app/(spaces)/SpacePage";
-import { SpaceConfig } from "@/app/(spaces)/Space";
-import TabBar from "@/common/components/organisms/TabBar";
-import {
-  RESOURCES_TAB_CONFIG,
-  NOUNS_TAB_CONFIG,
-  SOCIAL_TAB_CONFIG,
-  GOVERNANCE_TAB_CONFIG,
-  FUNDED_WORKS_TAB_CONFIG,
-} from "./homePageTabsConfig";
-import { INITIAL_SPACE_CONFIG_EMPTY } from "@/constants/initialPersonSpace";
-
-const getTabConfig = (tabName: string) => {
-  switch (tabName) {
-    case "Governance":
-      return GOVERNANCE_TAB_CONFIG;
-    case "Resources":
-      return RESOURCES_TAB_CONFIG;
-    case "Funded Works":
-      return FUNDED_WORKS_TAB_CONFIG;
-    case "Social":
-      return SOCIAL_TAB_CONFIG;
-    case "Nouns":
-      return NOUNS_TAB_CONFIG;
-    default:
-      return NOUNS_TAB_CONFIG;
-  }
-};
-
-const Home = () => {
-  const router = useRouter();
-  const params = useParams();
-  const { getIsAccountReady, getIsInitializing } = useAppStore((state) => ({
-    getIsAccountReady: state.getIsAccountReady,
-    getIsInitializing: state.getIsInitializing,
-  }));
-  const isLoggedIn = getIsAccountReady();
-  const isInitializing = getIsInitializing();
-
-  // Local state to manage current tab name and ordering
-  const tabOrdering = ["Nouns","Social", "Governance", "Resources", "Funded Works"];
-  const [tabName, setTabName] = useState<string>("Nouns");
-
-  useEffect(() => {
-    const newTabName = params?.tabname
-      ? decodeURIComponent(params.tabname as string)
-      : "Nouns";
-
-    setTabName(newTabName);
-  }, []);
-
-  function switchTabTo(newTabName: string) {
-    router.push(`/home/${newTabName}`);
-  }
-
-  const tabBar = (
-    <TabBar
-      getSpacePageUrl={(tab) => `/home/${tab}`}
-      inHomebase={false}
-      currentTab={tabName}
-      tabList={tabOrdering}
-      switchTabTo={switchTabTo}
-      inEditMode={false}
-      updateTabOrder={async () => Promise.resolve()}
-      deleteTab={async () => Promise.resolve()}
-      createTab={async () => Promise.resolve({ tabName })}
-      renameTab={async () => Promise.resolve({ tabName })}
-      commitTab={async () => Promise.resolve()}
-      commitTabOrder={async () => Promise.resolve()}
-    />
-  );
-
-  const args: SpacePageArgs = isInitializing
-    ? {
-        config: { ...INITIAL_SPACE_CONFIG_EMPTY, isEditable: false } as SpaceConfig,
-        saveConfig: async () => {},
-        commitConfig: async () => {},
-        resetConfig: async () => {},
-        tabBar: tabBar,
-      }
-    : !isLoggedIn
-      ? {
-          config: getTabConfig(tabName) as SpaceConfig,
-          saveConfig: async () => {},
-          commitConfig: async () => {},
-          resetConfig: async () => {},
-          tabBar: tabBar,
-        }
-      : {
-          config: getTabConfig(tabName) as SpaceConfig,
-          saveConfig: async () => {},
-          commitConfig: async () => {},
-          resetConfig: async () => {},
-          tabBar: tabBar,
-        };
-
-  return <SpacePage key={tabName} {...args} />;
-};
-
-export default Home;
+export default HomeWrapper;
